fixed bug if use_maximum_filter=False

+++ NuRadioReco/modules/neutrinoVertexReconstructor/neutrino3DVertexReconstructor.py
@@ -736,7 +736,7 @@ class neutrino3DVertexReconstructor:
                             )
                     res[i_x_0:i_x_1, i_y_0:i_y_1] = np.take(maximized_correlation, corr_index[i_x_0:i_x_1, i_y_0:i_y_1].astype(int))
         else:
-            res = np.abs(self.__correlation)[corr_index]
+            res = np.abs(self.__correlation)[corr_index.astype(int)]
         res[~mask] = 0
         """
         corr_index = self.__correlation.shape[0] / 2 + np.round(delta_t * self.__sampling_rate)
